debugging

+++ src/deploy/Linux/noobaa_service_installer.sh
@@ -5,11 +5,13 @@ echo "installing NooBaa"
 PATH=/usr/local/noobaa:$PATH;
 mkdir /usr/local/noobaa/logs
 
-/usr/local/noobaa/node_modules/forever-service/bin/forever-service delete noobaa_local_service
+/usr/local/noobaa/node_modules/forever-service/bin/forever-service delete noobaa_local_service &> /dev/null
 
 if [ -f /usr/bin/systemctl ] || [ -f /bin/systemctl ]; then
   echo "Systemd detected. Installing service"
-  rm /etc/systemd/system/multi-user.target.wants/noobaalocalservice.service
+  if [ -f /etc/systemd/system/multi-user.target.wants/noobaalocalservice.service]; then
+    rm /etc/systemd/system/multi-user.target.wants/noobaalocalservice.service
+  fi
   cp /usr/local/noobaa/src/agent/system_d.conf /etc/systemd/system/multi-user.target.wants/noobaalocalservice.service
   systemctl enable noobaalocalservice
   systemctl daemon-reload
@@ -17,17 +19,12 @@ if [ -f /usr/bin/systemctl ] || [ -f /bin/systemctl ]; then
 elif [[ -d /etc/init ]]; then
   echo "Upstart detected. Creating startup script"
   if [ -f /etc/init/noobaalocalservice.conf ]; then
-    echo "Service already installed. Removing old service"
-    initctl stop noobaalocalservice
     rm /etc/init/noobaalocalservice.conf
   fi
   cp /usr/local/noobaa/src/agent/upstart.conf /etc/init/noobaalocalservice.conf
-  sleep 2
   initctl start noobaalocalservice
 elif [[ -d /etc/init.d ]]; then
   echo "System V detected. Installing service"
-#  /usr/local/noobaa/node /usr/local/noobaa/src/agent/agent_linux_installer --uninstall
-#  sleep 5
   /usr/local/noobaa/node /usr/local/noobaa/src/agent/agent_linux_installer
   type chkconfig &> /dev/null
   if [ $? -eq 0 ]; then
@@ -35,17 +32,17 @@ elif [[ -d /etc/init.d ]]; then
   else
     update-rc.d noobaalocalservice enable
   fi
+  service noobaalocalservice start
 else
   echo "ERROR: Cannot detect init mechanism! Attempting to force service installation"
-  if [ -f /etc/init/noobaalocalservice.conf ]; then
-    service noobaalocalservice stop
-    rm /etc/init/noobaalocalservice.conf
-  fi
-  /usr/local/noobaa/node /usr/local/noobaa/src/agent/agent_linux_installer --uninstall
   /usr/local/noobaa/node /usr/local/noobaa/src/agent/agent_linux_installer
   systemctl enable noobaalocalservice
   cp /usr/local/noobaa/src/agent/noobaalocalservice.conf /etc/init/noobaalocalservice.conf
-  service noobaalocalservice restart
+  if [ $? -eq 0 ]; then
+    chkconfig noobaalocalservice on
+  else
+    update-rc.d noobaalocalservice enable
+  fi
 fi
 
 echo "NooBaa installation complete"
